REF: move misplaced tests (#31189)
